Adds a `Table of Contents` to the `.md` for JULEP3
This document is very big, and I wasn't sure how to get a feel for what all the content was. I think a ToC helps a lot with that.

I autogenerated this Table of Contents by copy-pasting the content of `Pkg3.md` into https://tableofcontents.herokuapp.com/.

Feel free to just say "no, this is silly" to this PR! :)